perf: enable edge caching by removing layout nonce dependency
- Remove headers() call from root layout that forced ALL pages dynamic
- Replace inline theme script with external public/theme-init.js
- Remove nonce props from SafeJsonLd and Analytics components
- Fix middleware: use public, s-maxage=86400 for MDX pages
- Add cache headers to next.config.ts for 9 public routes
- Split TreeGallery: extract ImageLightbox client component
- Fix CompareInToolButton: accept string|array species prop
- Add force-dynamic to 9 auth-dependent pages

1465 static pages now eligible for Vercel edge caching (24h TTL, 7d SWR).
Previously ALL pages were dynamically rendered on every request.
