+++ src/app/[locale]/admin/images/page.tsx
@@ -4,6 +4,8 @@ import type { Metadata } from "next";
 import { Link } from "@i18n/navigation";
 import ImageReviewClient from "./ImageReviewClient";
 
+export const dynamic = "force-dynamic";
+
 type Props = {
   params: Promise<{ locale: string }>;
 };
