+++ src/app/[locale]/contribute/page.tsx
@@ -3,6 +3,8 @@ import { getTranslations } from "next-intl/server";
 import { ContributeClient } from "./ContributeClient";
 import { allTrees, type Tree } from "contentlayer/generated";
 
+export const dynamic = "force-dynamic";
+
 interface PageProps {
   params: Promise<{ locale: string }>;
 }
