+++ src/app/[locale]/contribute/photo/page.tsx
@@ -3,6 +3,8 @@ import { allTrees } from "contentlayer/generated";
 import type { Metadata } from "next";
 import PhotoUploadClient from "./PhotoUploadClient";
 
+export const dynamic = "force-dynamic";
+
 type Props = {
   params: Promise<{ locale: string }>;
 };
